Pushing the Shiny table joins into the locally run Analysis.R  file
Disable the use of AnalysisAdList file for the model fitting for now.
Create exported versions of the 3 main failes( sessiontable, AdList,AdDetailList
Push the merges from the shiny app(app.R) to the Analysis.R
The SnapShatDate value had been modified to count as the prior's daay if the hour() part is below <11 hours.
Added a remove duplicate part to avoid double counting the volume when multiple scans were run on the same say.
An Additional criteria had been implemented to filter out scans where the receivednumber and resultnumbers were below the expected 15'000